backend/requirements.txt
arrow==1.2.2
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.2
cffi==1.15.0
Django==4.0.5
django-cors-headers==3.13.0
dnspython==2.2.1
email-validator==1.2.1
fqdn==1.5.1
idna==3.3
isoduration==20.11.0
jsonpointer==2.3
jsonschema==4.6.0
Pillow==9.1.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
pyzbar==0.1.9
qrcode==7.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
uri-template==1.2.0
webcolors==1.12
